\name{validate_spec}
\alias{validate_spec}
\title{Validate a PMSLT model specification}
\usage{
validate_spec(spec)
}
\arguments{
\item{spec}{A \code{pmslt_spec} object.}
}
\value{
Invisibly returns \code{TRUE} if valid.
}
\description{
Validate a PMSLT model specification before generating templates or running
modelling workflow steps.
}
